Update distanceAccuracyList wherever applicable during the game play. The only place where it has not been updated yet is evaluateResult, because evaluateResult has not been implemented yet.

+++ Map Guessr/Services/CoreGameService.swift
@@ -73,7 +73,7 @@ class CoreGameService: ObservableObject {
         email: String,
         level: Int,
         guessesLeft: Int,
-        accuracyInKm: Float,
+        accuracyInKm: Int,
         timeLapseInGame: Int
     ) async -> DistanceResponse? {
         guard let url = URL(string: AppConfig.Endpoints.evaluate) else {
